delete file.txt

+++ kryuchkova.kira/S1/main.cpp
@@ -50,6 +50,12 @@ int main(int argc, char *argv[])
       return 1;
     }
   }
+  if (results.isEmpty())
+  {
+    std::cout << "Empty file!\n";
+    return 1;
+  }
+
   std::cout << results.drop();
   while (!results.isEmpty())
   {
